+++ snowddl/resolver/schema_role.py
@@ -54,6 +54,7 @@ class SchemaRoleResolver(AbstractRoleResolver):
             )
 
         ownership_object_types = [
+            ObjectType.DYNAMIC_TABLE,
             ObjectType.EXTERNAL_TABLE,
             ObjectType.FILE_FORMAT,
             ObjectType.FUNCTION,